reafactor(fuse): add read/write settings
