+++ src/query/service/tests/it/storages/fuse/operations/gc.rs
@@ -21,7 +21,6 @@ use common_catalog::table_context::TableContext;
 use common_datablocks::DataBlock;
 use common_exception::Result;
 use common_storages_fuse::io::write_meta;
-use common_storages_fuse::io::BlockWriter;
 use common_storages_fuse::io::SegmentWriter;
 use common_storages_fuse::statistics::gen_columns_statistics;
 use common_storages_fuse::FuseTable;
@@ -33,6 +32,7 @@ use common_storages_table_meta::meta::Versioned;
 use futures_util::TryStreamExt;
 use uuid::Uuid;
 
+use crate::storages::fuse::block_writer::BlockWriter;
 use crate::storages::fuse::table_test_fixture::append_sample_data;
 use crate::storages::fuse::table_test_fixture::check_data_dir;
 use crate::storages::fuse::table_test_fixture::TestFixture;
@@ -246,6 +246,7 @@ mod utils {
 
     use chrono::DateTime;
     use chrono::Utc;
+    use common_storages_fuse::FuseStorageFormat;
 
     use super::*;
 
@@ -301,12 +302,7 @@ mod utils {
         for block in blocks {
             let stats = gen_columns_statistics(&block, None)?;
             let block_meta = block_writer
-                .write(
-                    common_storages_fuse::FuseStorageFormat::Parquet,
-                    block,
-                    stats,
-                    None,
-                )
+                .write(FuseStorageFormat::Parquet, block, stats, None)
                 .await?;
             block_metas.push(Arc::new(block_meta));
         }
