+++ src/query/storages/fuse/src/io/mod.rs
@@ -25,6 +25,7 @@ pub use read::load_bloom_filter_by_columns;
 pub use read::BlockFilterReader;
 pub use read::BlockReader;
 pub use read::MetaReaders;
+pub use read::ReadSettings;
 pub use read::SegmentInfoReader;
 pub use read::SnapshotHistoryReader;
 pub use read::TableSnapshotReader;
@@ -36,5 +37,5 @@ pub use snapshots::SnapshotsIO;
 pub use write::write_block;
 pub use write::write_data;
 pub use write::write_meta;
-pub use write::BlockWriter;
 pub use write::SegmentWriter;
+pub use write::WriteSettings;
